+++ packages/design-system/components/native/tabs/Tab/styles.ts
@@ -5,6 +5,7 @@ import { Theme } from '../../../../styles/native'
 export const createStyles = (theme: Theme, isSelected: boolean) =>
   StyleSheet.create({
     baseTabContainer: {
+      flexGrow: 1,
       borderBottomColor: theme.colors.object.low,
       borderBottomWidth: isSelected ? 2 : 0,
     },