+++ packages/wagtail/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @baseapp-frontend/wagtail
 
+## 1.0.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.1.8
+
+
 ## 1.0.44
 
 ### Patch Changes